requirements.txt
flask
torch
torchvision
pillow
flask_sqlalchemy
